RHEL 7: tests/test-partition1.sh: Skip this test if sfdisk has no -X option.
Fixes commit 21185fe4f29ba2b3ca4de12038023824dbea420e.

(cherry picked from commit b4217556ef02e016505da8be9ae619abd8b3ed68)

+++ tests/test-partition1.sh
@@ -38,6 +38,12 @@ requires sfdisk --help
 requires test -r /dev/urandom
 requires qemu-img --version
 
+# RHEL 7 sfdisk didn't have the -X option, so skip the tests here.
+if LANG=C sfdisk -X |& grep -sq "invalid option"; then
+    echo "$0: skipping test because no sfdisk -X option"
+    exit 77
+fi
+
 d="partition1.d"
 rm -rf $d
 cleanup_fn rm -rf $d
